ReadReq: Move stuff into common file
Read request handling is getting fairly large
